build: 1/2 remove squisher
build: Pass OTAPACKAGE location to squisher

Pass the OTAPACKAGE location so we don't have to guess what
the package name will be. This fixes an issue when
BUILD_NUMBER is defined in an environment with automated
builds.

Change-Id: Iab18b541de37c99d237aabc3cb7b2e9ab076617b

Fix building when OUT_DIR is out of ANDROID_BUILD_TOP

This broke in http://review.cyanogenmod.com/#/c/23686/ / http://review.cyanogenmod.com/#/c/23687/

Change-Id: I9dc77a1fb465a30c833262ba3b1de3d2f0f968f0

envsetup: set OUT_DIR to an absolute path always

OUT_DIR was set to $(TOPDIR)out previously,
but $(TOPDIR) was null, so it was a relative path.
This broke releasetools, inline kernel building, etc
since they require absolute paths.
Fix it so that it is set to $(shell readlink -f .)/out
if $(TOPDIR) is null.

Also remove hacks which checked if (OUT_DIR) was out
and changed it to $(ANDROID_BUILD_TOP)/out to workaround
the aforementioned problem.

Change-Id: I459a3b1325a1bbea0565cd73f6acf160d4ed9b39