gettop: fix symlink detection
Summary:
We must use the -P flag for pwd to properly get $PWD from the environment *with* symlinks. Otherwise symlink
detection is broken and anything we do using $(gettop) at the top of a symlinked path won't work properly. I noticed this
issue while running on a system that has a SSD RAID symlinked to my android build home directory, which is ~/ssd_storage.
The top in this instance is a symlink, which means that gettop() returns my home path, or the path above the symlink
directory (but not the symlinked directory). This is a major issue in case developers choose to use $(gettop),
$ANDROID_BUILD_TOP, or $T (doesn't really matter either way).

Test Plan:
Create a symlink and sync android source into the path. Open up the symlink, run envsetup.sh, and check
$(gettop) output.

Reviewers: Mustaavalkosta, michyprima, rlasalle15, randall.rushing

Differential Revision: http://review.codefirex.com/D188

Change-Id: Ide568d04d2a1c7d927126203f192ec1178b0f65f